test: move restore failure gate coverage into restore test
